Move serializer into seperate files
